Add tests for sync events
Didn't have time to write tests for these in ad60504.

These tests are based on JS's current behaviour, which was not tested;
I've implemented the equivalent tests for JS in [1]. Will use these
tests as a basis for writing the spec in #80.

To get these tests to pass I've had to add two behaviours from JS that
I missed in ad60504:

1. Don't always transition to SYNCING on ATTACHED
2. Transition to SYNCING when an OBJECT_SYNC is received when already
   SYNCED

[1] https://github.com/ably/ably-js/pull/2121