Use the sync state to drive getRoot's wait-for-sync
The `state` variable was added in ad60504 alongside the existing
`syncStatus` variable, introducing redundant concepts of "is the sync
completed?". Use the `state` as our source of truth, to be consistent
with JS and eliminate the redundancy.

In #80, I'll update the spec to make this behaviour explicit.

Note that this commit looks more complicated than it should; this is
because (as commented in the code) adding a new internal
SubscriptionStorage is much harder than it should be; will address this
in #102.